+++ main.lua
@@ -1,0 +1,9 @@
+------------------------
+-- Project Small R 
+-- Author: SJoshua
+------------------------
+logger = require("logger")
+config = require("config")
+bot = require("bot")
+
+bot.run()
