+++ old/bot.lua
@@ -1,0 +1,12 @@
+-------------------------------------------
+-- Project Small R
+-- Main
+-- Au: SJoshua
+-------------------------------------------
+dofile("api.lua")
+dofile("soul.lua")
+dofile("config.lua")
+
+print("Project Small R: Link Start.")
+
+bot.run()
