+++ old/config-example.lua
@@ -1,0 +1,17 @@
+-------------------------------------------
+-- Project Small R
+-- Example of Configure
+-- Au: SJoshua
+-------------------------------------------
+config = {
+    token = "BOT_TOKEN",
+	master = "MASTER_USERNAME",
+	masterid = 0,
+	dolua = true,
+	mute = false,
+	warning = false,
+	debug = false,
+	timeout = 0,
+	limit = 100
+}
+
