+++ requirements.txt
@@ -1,0 +1,3 @@
+lua-cjson
+LuaLogging
+multipart-post
